scripts/requirements.txt
open-clip-torch>=2.20.0
insightface>=0.7.3
onnxruntime-rocm
hdbscan>=0.8.29
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=9.0.0
tqdm>=4.65.0
